[TRL-457] chore: 일정 이동 관련 예외 -> common 패키지 이동
