+++ src/main/java/com/cosain/trilo/trip/application/day/service/day_color_update/DayColorUpdateService.java
@@ -1,6 +1,6 @@
 package com.cosain.trilo.trip.application.day.service.day_color_update;
 
-import com.cosain.trilo.trip.application.exception.DayNotFoundException;
+import com.cosain.trilo.common.exception.day.DayNotFoundException;
 import com.cosain.trilo.trip.application.exception.NoDayUpdateAuthorityException;
 import com.cosain.trilo.trip.domain.entity.Day;
 import com.cosain.trilo.trip.domain.repository.DayRepository;
