+++ src/main/java/com/cosain/trilo/trip/application/schedule/service/schedule_delete/ScheduleDeleteService.java
@@ -1,7 +1,7 @@
 package com.cosain.trilo.trip.application.schedule.service.schedule_delete;
 
+import com.cosain.trilo.common.exception.schedule.ScheduleNotFoundException;
 import com.cosain.trilo.trip.application.exception.NoScheduleDeleteAuthorityException;
-import com.cosain.trilo.trip.application.exception.ScheduleNotFoundException;
 import com.cosain.trilo.trip.domain.entity.Schedule;
 import com.cosain.trilo.trip.domain.repository.ScheduleRepository;
 import lombok.RequiredArgsConstructor;
