+++ src/main/java/com/cosain/trilo/trip/application/trip/service/temporary_search/TemporarySearchService.java
@@ -1,9 +1,9 @@
 package com.cosain.trilo.trip.application.trip.service.temporary_search;
 
+import com.cosain.trilo.common.exception.schedule.ScheduleNotFoundException;
+import com.cosain.trilo.common.exception.trip.TripNotFoundException;
 import com.cosain.trilo.trip.application.dao.ScheduleQueryDAO;
 import com.cosain.trilo.trip.application.dao.TripQueryDAO;
-import com.cosain.trilo.trip.application.exception.ScheduleNotFoundException;
-import com.cosain.trilo.common.exception.trip.TripNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
